attempting to fix app crashing

+++ app/src/main/java/com/example/wpms/MainActivity.kt
@@ -105,7 +105,7 @@ class MainActivity : AppCompatActivity() {
                     }
                     //insert pressure values into pressure table of DB
                     val pressureData = Pressure(
-                        pressureValue = pressureOneVal,
+                        pressureValue = pressureOneVal.toInt(),
 //                        timestamp = System.currentTimeMillis()
                     )
                     insertPressureDataIntoDB(pressureData)
